Quine.Beam

Description

Documentation

data Beam

A Beam is a fat Ray. It starts out with a given beamWidth and grows by beamWidthDelta every unit of time along the ray.

When used to model a focus the negative width is actually a real thing, so these do not stop when they hit width 0, but start to widen out again.

Constructors

Beam

Fields

_beamRay :: !Ray

_beamWidth :: !Float

_beamWidthDelta :: !Float

class ToRay t => ToBeam t where

Minimal complete definition

toBeam

Methods

toBeam :: t -> Beam

toBeamWidth :: t -> Float

toBeamWidthDelta :: t -> Float

class HasRay t => HasBeam t where

Minimal complete definition

beam

Methods

beam :: Lens' t Beam

beamWidth :: Lens' t Float

beamWidthDelta :: Lens' t Float

buildBeam :: (ToSphere a, ToSphere b) => a -> b -> Beam

evalBeam :: Beam -> Float -> (Vec3, Float)