Quine.Display

Description

Documentation

data Display

Basic OpenGL + SDL display that provides the screen and various bits of metadata about the act of displaying data

Constructors

Display

Fields

_displayWindow :: !Window

_displayGL :: !GLContext

_displayFullScreen :: !Bool

_displayWindowSize :: (Int, Int)

_displayWindowSizeChanged :: !Bool

_displayMinimized :: !Bool

_displayHasMouseFocus :: !Bool

_displayHasKeyboardFocus :: !Bool

_displayVisible :: !Bool

class HasDisplay c where

Minimal complete definition

display

Methods

display :: Lens' c Display

displayFullScreen :: Lens' c Bool

displayGL :: Lens' c GLContext

displayHasKeyboardFocus :: Lens' c Bool

displayHasMouseFocus :: Lens' c Bool

displayMinimized :: Lens' c Bool

displayVisible :: Lens' c Bool

displayWindow :: Lens' c Window

displayWindowSize :: Lens' c (Int, Int)

displayWindowSizeChanged :: Lens' c Bool

warn :: (MonadIO m, MonadState s m, HasDisplay s) => String -> String -> m ()

Complain loudly enough to pop up a window

resizeDisplay :: (MonadIO m, MonadReader e m, HasEnv e, MonadState s m, HasDisplay s) => m ()

Recalculate the actual OpenGL viewport size considering Retina

handleDisplayEvent :: (MonadIO m, MonadState s m, HasDisplay s, MonadReader e m, HasOptions e) => Event -> m ()

Discharge events we should always handle correctly, e.g. CUA concerns for quitting, going full-screen, etc.