Quine.GL

Description

Shader compilation environment

data ShaderEnv

class HasShaderEnv c where

Minimal complete definition

shaderEnv

Methods

shaderEnv :: Lens' c ShaderEnv

shaderEnvCpphsOpts :: Lens' c CpphsOptions

shaderEnvFragmentHighPrecisionAvailable :: Lens' c Bool

buildShaderEnv :: Options -> IO ShaderEnv

Compiling and Linking

compile :: (MonadIO m, MonadReader e m, HasShaderEnv e) => ShaderType -> FilePath -> m Shader

Compile a shader with a given set of defines

link :: MonadIO m => Shader -> Shader -> m Program

Link a program and vertex shader to build a program

Exceptions

data ProgramException

Constructors

ProgramException

Fields

programExceptionLog :: ByteString

_ProgramException :: Prism' SomeException ProgramException

data ShaderException

Constructors

ShaderException

Fields

shaderExceptionFileName :: String

shaderExceptionLog :: ByteString

_ShaderException :: Prism' SomeException ShaderException

Internals

cpp :: (MonadIO m, MonadReader e m, HasShaderEnv e) => FilePath -> m String

C preprocess a file

defined :: HasShaderEnv s => Lens' s [(String, String)]