Quine.GL.Attribute

Description

OpenGL Attribute

Documentation

class Attribute a where

Minimal complete definition

components, baseType

Methods

components :: p a -> Int

1, 2, 3 or 4 supported

baseType :: p a -> BaseType

normalize :: p a -> Bool

specifies whether integer data values should be normalized (True) or converted directly as float values (False) when they are accessed

Attribute Location

type AttributeLocation = GLuint

attributeLocation :: MonadIO m => Program -> String -> m (Maybe AttributeLocation)

Returns Nothing if the attribute is unbound within the program.

Attribute Definition

vertexAttribute :: HasLayoutAnnotation a => AttributeLocation -> SettableStateVar (Maybe (AttributeAccessor a b))

Associates the vertex attribute to the data in the vertex buffer. The association is stored in the vertex array object OpenGL wise, so a VAO must be bound

Attribute Pointer

vertexAttribPointerI :: MonadIO m => AttributeLocation -> Layout -> m ()

A not so high level binding to glVertexAttribIPointer sets the attribute array pointer for an integer attribute (no conversion)

vertexAttribPointer :: MonadIO m => AttributeLocation -> Layout -> m ()

A not so high level binding to glVertexAttribPointer sets the attribute array pointer for an integer or floating attribute (integers are converted to the floating type)

Layout

data Layout

Constructors

Layout Components BaseType Normalized Stride OffsetPtr

type BaseType = GLenum

type Components = Int

type Normalized = Bool

type Stride = Int

type OffsetPtr = Ptr ()

Layout Annotation

newtype UnAnnotated a

Constructors

UnAnnotated

Fields

unAnnotate :: a

newtype LayoutAnnotation a

Constructors

LayoutAnnotation

Fields

getLayout :: Layout

class HasLayoutAnnotation a where

Methods

layoutAnnotation :: Functor p => p (a UnAnnotated) -> a LayoutAnnotation

Annotates any proxy p of an attribute a with memory Layout informations

data AsType

Constructors

AsInteger

AsFloating