Quine.GL.Block

Description

OpenGL STD140 and STD430 support

Note STD430 can only be used for shader storage blocks, NOT uniform blocks!

Documentation

newtype Offset a b

Constructors

Offset Int

class Block b where

This describes how to load and store primitives through a uniform/shader storage blocks according to OpenGL STD140 and STD430.

There are lots of fiddly little constants around, beware.

Minimal complete definition

Nothing

Methods

alignment140 :: p b -> Int

As per Storable alignment, but matching OpenGL STD140.

sizeOf140 :: p b -> Int

As per Storable sizeOf, but matching OpenGL STD140.

isStruct :: p b -> Bool

Structures get smashed up to a minimum of a vec4 alignment in 140 mode

read140 :: MonadIO m => Ptr a -> Offset a b -> m b

write140 :: MonadIO m => Ptr a -> Offset a b -> b -> m ()

alignment430 :: p b -> Int

As per Storable alignment, but matching OpenGL STD430.

sizeOf430 :: p b -> Int

As per Storable sizeOf, but matching OpenGL STD430.

read430 :: MonadIO m => Ptr a -> Offset a b -> m b

write430 :: MonadIO m => Ptr a -> Offset a b -> b -> m ()

class GBlock f where

Automatically derive STD140 and STD430 alignment using GHC Generics

Methods

galignment140 :: p f -> Int

galignment430 :: p f -> Int

gsizeOf140 :: p f -> Int

gsizeOf430 :: p f -> Int

gread140 :: Ptr a -> Int -> IO (f b)

gread430 :: Ptr a -> Int -> IO (f b)

gwrite140 :: Ptr a -> Int -> f b -> IO ()

gwrite430 :: Ptr a -> Int -> f b -> IO ()