Quine.GL.Buffer

Description

OpenGL Doc: https://www.opengl.org/sdk/docs/man/html/glBindBuffer.xhtml

Also usable for bindless rendering: https://www.opengl.org/discussion_boards/showthread.php/170388-Bindless-Stuff

Documentation

newtype Buffer a

A Buffer is the generic OpenGL storage object for multiple possible kind of data

For ArrayBuffer it storages vertex attributes like position, normal or color an provides the MD (Multiple Data) in SIMD (Single Instruction, Multiple Data)

Constructors

Buffer GLuint

boundBufferAt :: BufferTarget -> StateVar (Buffer a)

Buffer Data

class BufferData a where

I bet there is already a package with

Methods

withRawData :: a -> (Int -> Ptr () -> IO ()) -> IO ()

perfom a monadic action with the pointer to the raw content and the size of it in bytes

fromRawData :: Int -> Ptr () -> IO a

bufferData :: forall a. BufferData a => BufferTarget -> StateVar (BufferUsage, a)

bufferDataDirect :: forall a. BufferData a => Buffer a -> StateVar (BufferUsage, a)

bindless uploading data to the argumented buffer (since OpenGL 4.5)

Buffer Targets

pattern ArrayBuffer :: BufferTarget

Vertex attributes

pattern DrawIndirectBuffer :: BufferTarget

Indirect command arguments

pattern ElementArrayBuffer :: BufferTarget

Vertex array indices

pattern PixelPackBuffer :: BufferTarget

Pixel read target

pattern PixelUnpackBuffer :: BufferTarget

Texture data source

pattern TransformFeedbackBuffer :: BufferTarget

Transform feedback buffer

pattern UniformBuffer :: BufferTarget

Uniform block storage

Buffer Usage

data BufferUsage

Usage Types

Terminology:

Stream

The data store contents will be modified once and used at most a few times.

Static

The data store contents will be modified once and used many times.

Dynamic

The data store contents will be modified repeatedly and used many times.

Draw

The data store contents are modified by the application, and used as the source for GL drawing and image specification commands.

Read

The data store contents are modified by reading data from the GL, and used to return that data when queried by the application.

Copy

The data store contents are modified by reading data from the GL, and used as the source for GL drawing and image specification commands.

pattern StreamDraw :: BufferUsage

pattern StreamRead :: BufferUsage

pattern StreamCopy :: BufferUsage

pattern StaticDraw :: BufferUsage

pattern StaticRead :: BufferUsage

pattern StaticCopy :: BufferUsage

pattern DynamicDraw :: BufferUsage

pattern DynamicRead :: BufferUsage

pattern DynamicCopy :: BufferUsage