Quine.GL.Error

Description

Documentation

newtype Error

Used to represent the result of glGetError

Constructors

Error GLenum

pattern NoError :: Error

No error has been recorded.

pattern InvalidEnum :: Error

An unacceptable value is specified for an enumerated argument.

pattern InvalidValue :: Error

A numeric argument is out of range.

pattern InvalidOperation :: Error

The specified operation is not allowed in the current state.

pattern InvalidFramebufferOperation :: Error

The framebuffer object is not complete.

pattern OutOfMemory :: Error

There is not enough memory left to execute the command.

pattern StackOverflow :: Error

This command would cause a stack overflow.

pattern StackUnderflow :: Error

This command would cause a stack underflow.

pattern TableTooLarge :: Error

The specified table exceeds the implementation's maximum supported table size.

Used by Graphics.GL.Raw.Exension.ARB.Imaging and Graphics.GL.Raw.EXT.Histogram

showError :: Error -> String

Generate a user-friendly version of an OpenGL error.

errors :: MonadIO m => m [Error]

No error in the result list should be NoError!

throwErrors :: MonadIO m => m ()

Throw Errors if we have any

newtype Errors

Constructors

Errors [Error]