Quine.GL.Pixel

Documentation

type InternalFormat = GLint

type PixelFormat = GLenum

type PixelType = GLenum

compSize1D :: PixelFormat -> PixelType -> Int -> Int

compSize1D pixelFormat pixelType n computes the size of a 1D texture of size n in bytes

compSize2D :: Int -> PixelFormat -> PixelType -> Int -> Int -> Int

compSize2D alignment pixelFormat pixelType w h computes the size of a 2D texture of size w * h in bytes where the rows are aligned to a given alignment = { 1, 2, 4, 8 } bytes.

compSize3D :: Int -> PixelFormat -> PixelType -> Int -> Int -> Int -> Int

compSize3D alignment pixelFormat pixelType w h d computes the size of a 3D texture of size w * h * d in bytes where the rows are aligned to a given alignment = { 1, 2, 4, 8 } bytes.