Quine.GL.Program

Description

Documentation

newtype Program

Constructors

Program GLuint

attachShader :: MonadIO m => Program -> Shader -> m ()

detachShader :: MonadIO m => Program -> Shader -> m ()

attachedShaders :: MonadIO m => Program -> m [Shader]

numAttachedShaders :: MonadIO m => Program -> m Int

numAttachedShaders program returns the number of shader objects attached to program.

linkProgram :: MonadIO m => Program -> m ()

linkStatus :: MonadIO m => Program -> m Bool

linkStatus program returns Trueif the last link operation on program was successful, False otherwise.

validateProgram :: MonadIO m => Program -> m ()

validateProgram program checks to see whether the executables contained in program can execute given the current OpenGL state. The information generated by the validation process will be stored in program's information log. The validation information may consist of an empty string, or it may be a string containing information about how the current program object interacts with the rest of current OpenGL state. This provides a way for OpenGL implementers to convey more information about why the current program is inefficient, suboptimal, failing to execute, and so on.

validateStatus :: MonadIO m => Program -> m Bool

validateStatus program returns True if the last validation operation on program was successful, and False otherwise.

If True, program is guaranteed to execute given the current state. Otherwise, program is guaranteed to not execute.

programInfoLog :: MonadIO m => Program -> m ByteString

programIsDeleted :: MonadIO m => Program -> m Bool

programIsDeleted program returns True if program is currently flagged for deletion, False otherwise.

numActiveAttributes :: MonadIO m => Program -> m Int

numActiveAttributes program returns the number of active attribute variables for program.

activeAttributeMaxLength :: MonadIO m => Program -> m Int

activeAttributeMaxLength program returns the length of the longest active attribute name for program, including the null termination character (i.e., the size of the character buffer required to store the longest attribute name). If no active attributes exist, 0 is returned.

numActiveUniforms :: MonadIO m => Program -> m Int

numActiveUniforms returns the number of active uniform variables for program@.

activeUniformMaxLength :: MonadIO m => Program -> m Int

activeUniformMaxLength program returns the length of the longest active uniform variable name for program, including the null termination character (i.e., the size of the character buffer required to store the longest uniform variable name). If no active uniform variables exist, 0 is returned.

activeAtomicCounterBuffers :: MonadIO m => Program -> m Int

activeAtomicCounterBuffers program returns the number of active attribute atomic counter buffers used by program.

programBinaryLength :: MonadIO m => Program -> m Int

programBinaryLength program return the length of the program binary, in bytes, that will be returned by a call to glGetProgramBinary. When a progam's linkStatus is False, its program binary length is 0.

programComputeWorkGroupSize :: MonadIO m => Program -> m (Int, Int, Int)

programComputeWorkgroupSize program returns three integers containing the local work group size of the compute program as specified by its input layout qualifier(s). program must be the name of a program object that has been previously linked successfully and contains a binary for the compute shader stage.

transformFeedbackVaryingsMaxLength :: MonadIO m => Program -> m Int

transformFeedbackVaryingsMaxLength program returns the length of the longest variable name to be used for transform feedback, including the null-terminator.

transformFeedbackBufferMode :: MonadIO m => Program -> m GLenum

transformFeedbackBufferMode program returns a symbolic constant indicating the buffer mode for program used when transform feedback is active. This may be GL_SEPARATE_ATTRIBS or GL_INTERLEAVED_ATTRIBS.

numTransformFeedbackVaryings :: MonadIO m => Program -> m Int

numTransformFeedbackVaryings program returns the number of varying variables to capture in transform feedback mode for the program.

geometryVerticesOut :: MonadIO m => Program -> m Int

geometryVerticesOut program returns the maximum number of vertices that the geometry shader in program will output.

geometryInputType :: MonadIO m => Program -> m GLenum

geometryInputType program returns a symbolic constant indicating the primitive type accepted as input to the geometry shader contained in program.

geometryOutputType :: MonadIO m => Program -> m GLenum

geometryOutputType program returns a symbolic constant indicating the primitive type that will be output by the geometry shader contained in program.

currentProgram :: StateVar Program