Quine.GL.Version

Documentation

vendor :: String

Returns the company responsible for this GL implementation. This name does not change from release to release.

renderer :: String

Returns the name of the renderer. This name is typically specific to a particular configuration of a hardware platform. It does not change from release to release.

version :: Version

Returns a version or release number.

shadingLanguageVersion :: Version

Return the primary shading language version

shadingLanguageVersions :: Set Version

Returns a set of shading language versions supported by this implementation.