Quine.Input

Description

Do we need to hook window entry to get a better state summary of the mouse buttons to avoid the common 'sticky mouse' problems in games?

Should losing focus kill the mouse button states?

Documentation

data Input

Constructors

Input

Fields

_keyCodes :: Set Int32

_scanCodes :: Set Word32

_mouseX :: Int32

_mouseY :: Int32

_mouseButtons :: Set Word8

_mouseButtonMask :: Word32

_mouseRelX :: Int32

_mouseRelY :: Int32

_mouseWheelX :: Int32

_mouseWheelY :: Int32

class HasInput c where

Minimal complete definition

input

Methods

input :: Lens' c Input

keyCodes :: Lens' c (Set Int32)

mouseButtonMask :: Lens' c Word32

mouseButtons :: Lens' c (Set Word8)

mouseRelX :: Lens' c Int32

mouseRelY :: Lens' c Int32

mouseWheelX :: Lens' c Int32

mouseWheelY :: Lens' c Int32

mouseX :: Lens' c Int32

mouseY :: Lens' c Int32

scanCodes :: Lens' c (Set Word32)

handleInputEvent :: (MonadState s m, HasInput s) => Event -> m ()