Quine.Math

Pointwise Manipulation

smoothstep :: (Fractional a, Ord a) => a -> a -> a -> a

smoothstep l h x goes from 0 to 1 over the interval from l to h

clamp :: Ord a => a -> a -> a -> a

clamp x l h clamps x to the range l <= x <= h

saturate :: (Num a, Ord a) => a -> a

saturate x returns x saturated to the range [0,1]

saturate x ≡ clamp x 0 1

mix :: Num a => a -> a -> a -> a

mix x y a linear interpolates between x and y, weighting y with weight a

step :: (Num a, Ord a) => a -> a -> a

step edge x is 0 for x < edge, 1 thereafter.

Vector Manipulation

reflect :: (Applicative f, Metric f, Num a) => f a -> f a -> f a

reflect i n eta calculates the reflection direction for an incident vector i around a normal n. i and n should be normalized.

refract :: (Applicative f, Metric f, Floating a, Ord a) => f a -> f a -> a -> f a

reflect i n eta calculates the refraction vector for an incident vector i around a normal n. i and n should be normalized.

faceforward :: (Applicative f, Metric f, Num a, Ord a) => f a -> f a -> f a -> f a

faceforward n i nref is used to return a vector pointing in the same direction as another.

It orients a vector to point away from the surface as defined by its normal.

n is the vector to orient (perturbed normal vector) i is the incidence vector, (e.g. vector from the eye) nref is the reference vector or geometric normal vector.

Color space and Tone mapping

luminance :: (Fractional a, Epsilon a) => Lens' (V3 a) a

Manipulate luminance in an RGB value. Note: when changing from Y = 0, we construct a greyscale value as the other axes are lost.

exposureAdjustment :: Num a => a

standardSRGB :: (Floating a, Ord a) => a -> a

Convert from linear to sRGB per the gl_EXT_texture_sRGB_decode extension

standardLinear :: (Floating a, Ord a) => a -> a

Convert from sRGB to linear as per the gl_EXT_texture_sRGB_decode extension

simplifiedSRGB :: Floating a => a -> a

Convert from linear to sRGB using the simplified gamma curve.

simplifiedLinear :: Floating a => a -> a

Convert from sRGB to linear using the simplified gamma curve.

simplifiedReinhard :: Fractional a => a -> a

Simplified componentwise Reinhard tonemap operator, result is still in "linear" space but in [0..1]

You should really be applying this to luminance, or you'll get particular milky greys.

http://www.gdcvault.com/play/1012351/Uncharted-2-HDR

color & luminance %~ simplifiedReinhard

You should apply an exposure adjustment and need to map it to the monitor by applying (** (1/2.2))

fullSimplifiedReinhard :: Floating a => a -> a

Reinhard tonemap operator, maps to the monitor working space, no adjustment required.

http://filmicgames.com/archives/75

enhancedReinhard :: Fractional a => a -> a -> a

Reinhard should really be applied to luminance values.

http://imdoingitwrong.wordpress.com/2010/08/19/why-reinhard-desaturates-my-blacks-3/

color & luminance %~ enhancedReinhard 4

enhancedReinhard l_white l takes the least color value that should be mapped to white and a color to map, and maps it.

The result remains in linear space.

fullEnhancedReinhard :: (Epsilon a, Floating a) => a -> V3 a -> V3 a

Takes a color all the way from linear to monitor working space

hejlBurgessDawson :: (Fractional a, Ord a) => a -> a

Jim Hejl and Richard Burgess-Dawson's tonemap, maps from linear into monitor space.

http://filmicgames.com/archives/75

No adjustment required.