Quine.Plane

Description

Assumes normals are actually unit vectors

Documentation

data Plane

Constructors

Plane

Fields

_planeNormal :: !Vec3

_planeDistance :: !Float

class ToNormal a => ToPlane a where

Methods

toPlane :: a -> Plane

class (HasNormal a, ToPlane a) => HasPlane a where

Methods

plane :: Lens' a Plane

distance :: Lens' a Float

signedDistance :: Plane -> Vec3 -> Float

pVertex :: Plane -> Box -> Vec3

this is the vertex of the box in the 'most positive' position relative to this plane.

Out of all corners of the box, this will compare with the greatest signed distance

nVertex :: Plane -> Box -> Vec3

this is the vertex of the box in the 'most negative' position relative to this plane.

Out of all corners of the box, this will compare with the least signed distance