Quine.Ref

Documentation

data Ref a

reference counting

Constructors

Ref (IO ()) a !(IORef Int)

newRef :: MonadIO m => IO () -> a -> m (Ref a)

create a reference counter initialized to 1

acquireRef :: MonadIO m => Ref a -> m a

Increase the reference count of a living ref by 1

releaseRef :: MonadIO m => Ref a -> m ()

Decrease the reference count of a living ref by 1, releasing it if it hits 0

validRef :: MonadIO m => Ref a -> m Bool

Check if the Ref has any references and can still be used

withRef :: MonadBaseControl IO m => Ref a -> (a -> m ()) -> m ()

Use a reference counted resource in a scope