Quine.Simulation

Description

simulation is a bit of a shell game, we need two physics states per rendering frame but we need two rendering frames for temporal antialiasing, so we'll have up to 4 physics states in flight at any given time.

Here we track the two current states

Documentation

data Simulation a

Constructors

Simulation

Fields

_simulationStart :: !Time

_simulationTime :: !Time

_simulationFrameCounter :: !Counter

_simulationOld :: !(Ref a)

_simulationState :: !(Ref a)

_simulationMeter :: !Meter

class HasSimulation c a | c -> a where

Minimal complete definition

simulation

Methods

simulation :: Lens' c (Simulation a)

simulationFrameCounter :: Lens' c Counter

simulationMeter :: Lens' c Meter

simulationOld :: Lens' c (Ref a)

simulationStart :: Lens' c Time

simulationState :: Lens' c (Ref a)

simulationTime :: Lens' c Time

simulationFPS :: Double

# of fps

simulationSPF :: DeltaTime

simulationBurstRate :: Int

Catch up rate, if we're at less than a frame every second, we're hurting!

createSimulation :: Simulated a => Monitor -> a -> a -> IO (Simulation a)

simulate :: (MonadIO m, MonadState s m, HasSimulation s a, Simulated a) => m () -> m (Double, Time)

class Simulated a where

Methods

newState :: MonadIO m => a -> a -> m a

deleteState :: MonadIO m => a -> m ()