Quine.Supply

Documentation

data Supply :: *

class HasSupply s where

Methods

supply :: Lens' s Supply

split :: (MonadState s m, HasSupply s) => m Supply

Split off a variable supply that will return distinct identifiers from those returned by the current supply.

fresh :: (MonadState s m, HasSupply s) => m Int

Get a globally unique fresh id, so long as every Supply is accessed by a single thread, and none leak reused keys.