Quine.Task

Description

The GPU is an asynchronous resource, but we only have one controlling thread..

We can poll the GPU at various sync points

Documentation

class MonadIO m => MonadTask m where

Methods

spawn :: m a -> m (m a)

Start a task, this task may be run out of order with your own context.

After running this command, assume that the context is scrambled.

After running the returned command, also assume that the context is scrambled.

fence :: m Fence

Create a synchronization point, later we can use the action we get to block

This does not block and does not scramble the context.

await :: Fence -> m ()

Assume the context is scrambled after this returns

barrier :: MonadTask m => m ()

Wait until all commands up to a given point have been completed

Meanwhile other tasks may run.

When you get control back, assume that the context is scrambled

barrier ≡ fence >>= wait

data Fence

Constructors

Fence !Int !GLsync

newtype Task a

Constructors

Task

Fields

runTask :: (a -> Continuation) -> Word64 -> IO () -> Continuation

execTask :: MonadIO m => Word64 -> IO () -> Task a -> m a

execTask delay idle task will run a task on the GPU and should only be run in the bound thread that is controlling the user interface.

When we block waiting on a sync from the GPU and have no ready fibers then we'll block for up to delay nanoseconds and if that timeout expires, we'll run idle