Quine.GL

Description

Shader compilation environment

data ShaderEnv

class HasShaderEnv c where

Minimal complete definition

shaderEnv

Methods

shaderEnv :: Lens' c ShaderEnv

shaderEnvCpphsOpts :: Lens' c CpphsOptions

shaderEnvFragmentHighPrecisionAvailable :: Lens' c Bool

buildShaderEnv :: Options -> IO ShaderEnv

Compiling and Linking

compile :: (MonadIO m, MonadReader e m, HasShaderEnv e) => ShaderType -> FilePath -> m Shader

Compile a shader with a given set of defines

link :: MonadIO m => Shader -> Shader -> m Program

Link a program and vertex shader to build a program

Exceptions

data ProgramException

Constructors

ProgramException

Fields

programExceptionLog :: String

_ProgramException :: Prism' SomeException ProgramException

data ShaderException

Constructors

ShaderException

Fields

shaderExceptionFileName :: String

shaderExceptionLog :: String

_ShaderException :: Prism' SomeException ShaderException

Internals

cpp :: (MonadIO m, MonadReader e m, HasShaderEnv e) => FilePath -> m String

C preprocess a file

defined :: HasShaderEnv s => Lens' s [(String, String)]

Caching Resources

data Cache a

Constructors

Cache

Fields

_cachePool :: [a]

_cacheSize :: Int

data Caches

Caches for several OpenGL resource types

Constructors

Caches

Fields

_textures :: !(Cache TextureObject)

_buffers :: !(Cache BufferObject)

_queries :: !(Cache QueryObject)

_renderbuffers :: !(Cache RenderbufferObject)

_vertexArrays :: !(Cache VertexArrayObject)

class HasCaches c where

Minimal complete definition

caches

Methods

caches :: Lens' c Caches

buffers :: Lens' c (Cache BufferObject)

queries :: Lens' c (Cache QueryObject)

renderbuffers :: Lens' c (Cache RenderbufferObject)

textures :: Lens' c (Cache TextureObject)

vertexArrays :: Lens' c (Cache VertexArrayObject)

class GeneratableObjectName a => Cached a where

Methods

cache :: HasCaches t => Lens' t (Cache a)

generate :: (MonadIO m, MonadState s m, HasCaches s, Cached a) => m a

Generate an OpenGL resource, leaning on a local a cache

generates :: (MonadIO m, MonadState s m, HasCaches s, Cached a) => Int -> m [a]

Generate many OpenGL resources, leaning on a local a cache

delete :: (MonadState s m, HasCaches s, Cached a) => a -> m ()

Return an OpenGL resource to the cache

purge :: (MonadState s m, MonadIO m, ObjectName a) => ALens' s (Cache a) -> m ()

Purge a cache

Exposed Internals

cachePool :: forall a a. Lens (Cache a) (Cache a) [a] [a]

cacheSize :: forall a. Lens' (Cache a) Int

State variables

(&=) :: (MonadIO m, HasSetter s) => s a -> a -> m () infixr 2

(&=!) :: (MonadIO m, HasSetter s) => s a -> a -> m () infixr 2

the :: MonadIO m => HasGetter s => s a -> m a

xmap :: (b -> a) -> (a -> b) -> StateVar a -> StateVar b