Quine.GL.Block

Description

OpenGL STD140 an STD430 support

Documentation

newtype Offset a b

Constructors

Offset Int

class Block b where

This describes how to load and store primitives through a uniform/shader storage blocks according to OpenGL STD140 and STD430.

There are lots of fiddly little constants around, beware.

Minimal complete definition

Nothing

Methods

alignment140 :: p b -> Int

As per Storable alignment, but matching OpenGL STD140.

sizeOf140 :: p b -> Int

As per Storable alignment, but matching OpenGL STD140.

isStruct :: p b -> Bool

read140 :: MonadIO m => Ptr a -> Offset a b -> m b

write140 :: MonadIO m => Ptr a -> Offset a b -> b -> m ()

alignment430 :: p b -> Int

As per Storable alignment, but matching OpenGL STD430.

sizeOf430 :: p b -> Int

As per Storable alignment, but matching OpenGL STD430.

read430 :: MonadIO m => Ptr a -> Offset a b -> m b

write430 :: MonadIO m => Ptr a -> Offset a b -> b -> m ()