Quine.Input

Description

Do we need to hook window entry to get a better state summary of the mouse buttons to avoid the common 'sticky mouse' problems in games?

Should losing focus kill the mouse button states?

Documentation

data Input

Constructors

Input

Fields

_keyCodes :: !(Set Int32)

_scanCodes :: !(Set Word32)

_mouseButtons :: !(Set Word8)

_mousePos :: !(V2 Int32)

_mouseButtonMask :: !Word32

_mouseRel :: !(V2 Int32)

_mouseWheel :: !(V2 Int32)

class HasInput c where

Minimal complete definition

input

Methods

input :: Lens' c Input

keyCodes :: Lens' c (Set Int32)

mouseButtonMask :: Lens' c Word32

mouseButtons :: Lens' c (Set Word8)

mousePos :: Lens' c (V2 Int32)

mouseRel :: Lens' c (V2 Int32)

mouseWheel :: Lens' c (V2 Int32)

scanCodes :: Lens' c (Set Word32)

handleInputEvent :: (MonadState s m, HasInput s) => Event -> m ()