Quine.Math

Pointwise Manipulation

smoothstep :: (Fractional a, Ord a) => a -> a -> a -> a

smoothstep l h x goes from 0 to 1 over the interval from l to h

clamp :: Ord a => a -> a -> a -> a

clamp x l h clamps x to the range l <= x <= h

saturate :: (Num a, Ord a) => a -> a

saturate x returns x saturated to the range [0,1]

saturate x ≡ clamp x 0 1

mix :: Num a => a -> a -> a -> a

mix x y a linear interpolates between x and y, weighting y with weight a

step :: (Num a, Ord a) => a -> a -> a

step edge x is 0 for x < edge, 1 thereafter.

Vector Manipulation

reflect :: (Applicative f, Metric f, Num a) => f a -> f a -> f a

reflect i n eta calculates the reflection direction for an incident vector i around a normal n. i and n should be normalized.

refract :: (Applicative f, Metric f, Floating a, Ord a) => f a -> f a -> a -> f a

reflect i n eta calculates the refraction vector for an incident vector i around a normal n. i and n should be normalized.

faceforward :: (Applicative f, Metric f, Num a, Ord a) => f a -> f a -> f a -> f a

faceforward n i nref is used to return a vector pointing in the same direction as another.

It orients a vector to point away from the surface as defined by its normal.

n is the vector to orient (perturbed normal vector) i is the incidence vector, (e.g. vector from the eye) nref is the reference vector or geometric normal vector.

Colorspace Manipulation

srgb :: (Floating a, Ord a) => a -> a

Convert from linear to sRGB

linear :: (Floating a, Ord a) => a -> a

Convert from sRGB to linear